enums: move files to thrift/thriftutil
