+++ thrift/thriftutil/enums_test.go
@@ -1,4 +1,4 @@
-package stringutil
+package thriftutil
 
 import (
 	"testing"
